use float

+++ USER/four_wheel.c
@@ -54,8 +54,8 @@ void f_move_xy_c(int8_t spd_x, int8_t spd_y)
 {
 	int16_t arg_x, arg_y;
 	
-	double coe_x = (sqrt( pow( CAR_X_LENGTH, 2) + pow( CAR_Y_LENGTH, 2))) / CAR_Y_LENGTH;
-	double coe_y = (sqrt( pow( CAR_X_LENGTH, 2) + pow( CAR_Y_LENGTH, 2))) / CAR_X_LENGTH;
+	float coe_x = (sqrtf( powf( CAR_X_LENGTH, 2) + powf( CAR_Y_LENGTH, 2))) / CAR_Y_LENGTH;
+	float coe_y = (sqrtf( powf( CAR_X_LENGTH, 2) + powf( CAR_Y_LENGTH, 2))) / CAR_X_LENGTH;
 
 	arg_x = (int16_t) (((float) spd_x) / 128 * DEFAULT_ARG_SPEED);
 	arg_y = (int16_t) (((float) spd_y) / 128 * DEFAULT_ARG_SPEED);
